SSARE/scraper_service/requirements.txt
fastapi
uvicorn
requests
bs4
pandas
aiohttp
sqlalchemy
prefect
prefect-ray
newspaper4k
lxml[html_clean]
pgvector
sqlmodel
redis
playwright
ray==2.39.0